% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials = 2, prob = 0.5)
}
\arguments{
\item{trials}{(integer) the total number of trials. Must be a positive integer.}

\item{prob}{(numeric) the probability of success on each trial. Must be between
0 and 1 inclusive.}
}
\value{
a data frame of class c("bincum", "data.frame") containing three columns:
success, probability, cumulative.
}
\description{
computes and tabulates the cumulative density for the given
binomial distribution. Also gives the density.
}
\examples{
bin_cumulative()
bin_distribution(trails = 2, prob = .1)
bin_distribution(trials = 6, prob = .3)
}
