% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_distribution.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials = 2, prob = 0.5)
}
\arguments{
\item{trials}{(integer) the number of trials}

\item{prob}{(real) the probability of success on each trial. Must be between 0 and 1.}
}
\value{
a data frame of class c("bindis", "dataframe") containing the probability distribution
}
\description{
Computes the probability distribution of a binomial with a given
number of trials and probability. It tabulates the probability of each number of
successes.
}
\examples{
bin_distribution()
bin_distribution(trails = 2, prob = .1)
bin_distribution(trials = 6, prob = .3)
}
