% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials. must be an integer}

\item{prob}{probability of success on each trial. Must be
between 0 and 1.}
}
\value{
a measure of the kurtosis of the given distribution.
}
\description{
computes the kurtosis of the given binomial distribution.
The kurtosis is a descriptor for the "tailness" of a distribution
}
\examples{
bin_kurtosis()
bin_kurtosis(5)
bin_kurtosis(100, .3)
}
