% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials. Must be a positive integer}

\item{prob}{probability of success on each trial. Must be
between 0 and 1}
}
\value{
the expected value of the given binomial distribution
}
\description{
computes the mean of a given binomial distribution
}
\examples{
bin_mean()
bin_mean(5)
bin_mean(100, .3)
}
