% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_measures.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials. Must be a positive integer.}

\item{prob}{probability of success in each trial. Must be
between 0 and 1}
}
\value{
the mode of the binomial distribution
}
\description{
computes the most likeley value(s) of the given
binomial distribution. i.e. the mode(s) of the distribution.
}
\examples{
bin_mode()
bin_mode(5)
bin_mode(100, .3)
}
