% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_probability.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(p = 0.5, trials = 1, success = 1)
}
\arguments{
\item{p}{(real) probability, 0 <= p <= 1. default .5}

\item{trials}{(integer) number of trials. Must be a positive integer}

\item{success}{(integer) number of desired successes, must be less than n
default 1}
}
\value{
the computed density of the binomial distribution for k
}
\description{
computes the probability density for a given
number of trials and successes, and a given probability.
}
\examples{
bin_probability(2, 5, .5)
bin_probability(success = 0:2, trials = 5, p = .45)
}
