% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials = 1, prob = 0.5)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability must be between 0 and 1 inclusive}
}
\value{
a named list containg the two parameters of the binomial random variable
}
\description{
creates an object of class binvar. A list with two named elements
trials and probability.
}
\examples{
bin_variable()
bin_variable(10, .3)
bin_variable(100, .75)
}
